+++ spec/mongoid/persistence/atomic_spec.rb
@@ -1,0 +1,5 @@
+require "spec_helper"
+
+describe Mongoid::Persistence::Atomic do
+
+end